ficx linters
